Project_Sugandh/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.0
django-base-url==2019.7.24
django-debug-toolbar==2.0
facebook-sdk==3.1.0
firebase==3.0.1
gcloud==0.18.3
googleapis-common-protos==1.6.0
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
jwcrypto==0.6.0
Naked==0.1.31
oauth2client==4.1.3
paramiko==2.6.0
protobuf==3.10.0rc1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pycryptodome==3.4.3
PyMySQL==0.9.3
PyNaCl==1.3.0
python-jwt==3.2.4
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
shellescape==3.4.1
six==1.12.0
sqlparse==0.3.0
sseclient==0.0.24
urllib3==1.25.3
whitenoise==4.1.3
